Add static functionality
Provide convenience functions for adding models, intensities,
and colour-maps when these objects exist pre-parsed. This
allows use without a server running. The ultimate goal of
these tweaks is to allow R to use brainbrowser as viewer
backend. R can handle the parsing, but it's convenient not
to require an active server.